--- docsource/api/selprotopy.SelClient.rst
@@ -1,60 +1,0 @@
-
-
-SelClient
-====================
-
-.. currentmodule:: selprotopy
-
-.. autoclass:: SelClient
-   :show-inheritance:
-
-   
-     
-   
-
-   
-   
-   
-
-   
-   
-
-   .. rubric:: Methods Summary
-
-   .. autosummary::
-   
-      ~SelClient.access_level
-      ~SelClient.access_level_1
-      ~SelClient.access_level_2
-      ~SelClient.autoconfig
-      ~SelClient.autoconfig_fastmeter
-      ~SelClient.autoconfig_fastmeter_demand
-      ~SelClient.autoconfig_fastmeter_peakdemand
-      ~SelClient.poll_fast_meter
-      ~SelClient.quit
-
-   
-   
-
-   
-   
-   
-
-   
-   
-
-   .. rubric:: Methods Documentation
-
-   
-   .. automethod:: access_level
-   .. automethod:: access_level_1
-   .. automethod:: access_level_2
-   .. automethod:: autoconfig
-   .. automethod:: autoconfig_fastmeter
-   .. automethod:: autoconfig_fastmeter_demand
-   .. automethod:: autoconfig_fastmeter_peakdemand
-   .. automethod:: poll_fast_meter
-   .. automethod:: quit
-
-   
-
