--- docsource/api/selprotopy.commands.event_record_request.rst
@@ -1,6 +1,0 @@
-selprotopy.commands.event\_record\_request
-==========================================
-
-.. currentmodule:: selprotopy.commands
-
-.. autofunction:: event_record_request
