--- docsource/api/selprotopy.commands.rst
@@ -1,20 +1,0 @@
-selprotopy.commands
-===================
-
-.. rubric:: Description
-
-.. automodule:: selprotopy.commands
-
-.. currentmodule:: selprotopy.commands
-
-
-
-
-.. rubric:: Functions
-
-.. autosummary::
-    :toctree: .
-    
-    event_record_request
-    
-
