--- docsource/api/selprotopy.common.ieee4bytefps.rst
@@ -1,6 +1,0 @@
-selprotopy.common.ieee4bytefps
-==============================
-
-.. currentmodule:: selprotopy.common
-
-.. autofunction:: ieee4bytefps
