--- docsource/api/selprotopy.common.int_to_bool_list.rst
@@ -1,6 +1,0 @@
-selprotopy.common.int\_to\_bool\_list
-=====================================
-
-.. currentmodule:: selprotopy.common
-
-.. autofunction:: int_to_bool_list
