--- docsource/api/selprotopy.common.rst
@@ -1,22 +1,0 @@
-selprotopy.common
-=================
-
-.. rubric:: Description
-
-.. automodule:: selprotopy.common
-
-.. currentmodule:: selprotopy.common
-
-
-
-
-.. rubric:: Functions
-
-.. autosummary::
-    :toctree: .
-    
-    ieee4bytefps
-    
-    int_to_bool_list
-    
-
