--- docsource/api/selprotopy.protoparser.FastMeterBlock.rst
@@ -1,6 +1,0 @@
-selprotopy.protoparser.FastMeterBlock
-=====================================
-
-.. currentmodule:: selprotopy.protoparser
-
-.. autofunction:: FastMeterBlock
