--- docsource/api/selprotopy.protoparser.FastMeterConfigurationBlock.rst
@@ -1,6 +1,0 @@
-selprotopy.protoparser.FastMeterConfigurationBlock
-==================================================
-
-.. currentmodule:: selprotopy.protoparser
-
-.. autofunction:: FastMeterConfigurationBlock
