--- docsource/api/selprotopy.protoparser.RelayBnaBlock.rst
@@ -1,6 +1,0 @@
-selprotopy.protoparser.RelayBnaBlock
-====================================
-
-.. currentmodule:: selprotopy.protoparser
-
-.. autofunction:: RelayBnaBlock
