--- docsource/api/selprotopy.protoparser.RelayDefinitionBlock.rst
@@ -1,6 +1,0 @@
-selprotopy.protoparser.RelayDefinitionBlock
-===========================================
-
-.. currentmodule:: selprotopy.protoparser
-
-.. autofunction:: RelayDefinitionBlock
