--- docsource/api/selprotopy.protoparser.RelayDnaBlock.rst
@@ -1,6 +1,0 @@
-selprotopy.protoparser.RelayDnaBlock
-====================================
-
-.. currentmodule:: selprotopy.protoparser
-
-.. autofunction:: RelayDnaBlock
