--- docsource/api/selprotopy.protoparser.RelayIdBlock.rst
@@ -1,6 +1,0 @@
-selprotopy.protoparser.RelayIdBlock
-===================================
-
-.. currentmodule:: selprotopy.protoparser
-
-.. autofunction:: RelayIdBlock
