--- docsource/api/selprotopy.protoparser.rst
@@ -1,30 +1,0 @@
-selprotopy.protoparser
-======================
-
-.. rubric:: Description
-
-.. automodule:: selprotopy.protoparser
-
-.. currentmodule:: selprotopy.protoparser
-
-
-
-
-.. rubric:: Functions
-
-.. autosummary::
-    :toctree: .
-    
-    FastMeterBlock
-    
-    FastMeterConfigurationBlock
-    
-    RelayBnaBlock
-    
-    RelayDefinitionBlock
-    
-    RelayDnaBlock
-    
-    RelayIdBlock
-    
-
