--- docsource/api/selprotopy.telnetlib_support.process_rawq.rst
@@ -1,6 +1,0 @@
-selprotopy.telnetlib\_support.process\_rawq
-===========================================
-
-.. currentmodule:: selprotopy.telnetlib_support
-
-.. autofunction:: process_rawq
