--- docsource/api/selprotopy.telnetlib_support.rst
@@ -1,20 +1,0 @@
-selprotopy.telnetlib\_support
-=============================
-
-.. rubric:: Description
-
-.. automodule:: selprotopy.telnetlib_support
-
-.. currentmodule:: selprotopy.telnetlib_support
-
-
-
-
-.. rubric:: Functions
-
-.. autosummary::
-    :toctree: .
-    
-    process_rawq
-    
-
