--- docsource/protoparser.rst
@@ -1,10 +1,0 @@
-.. _selprotopy.protoparser.py:
-
-.. rubric:: Modules
-
-.. autosummary::
-   :recursive:
-   :toctree: api
-   :template: module.rst
-
-   selprotopy.protoparser
